refactor: exported Row interface as separte file.
